move adapters into own package
