+++ apidef/adapter/gqlengineadapter/adapter_proxy_only.go
@@ -1,4 +1,4 @@
-package adapter
+package gqlengineadapter
 
 import (
 	"net/http"
@@ -10,19 +10,25 @@ import (
 	"github.com/TykTechnologies/tyk/apidef"
 )
 
-type proxyOnlyGraphQLEngineAdapter struct {
-	apiDefinition   *apidef.APIDefinition
-	schema          *graphql.Schema
-	httpClient      *http.Client
-	streamingClient *http.Client
+type ProxyOnly struct {
+	ApiDefinition   *apidef.APIDefinition
+	HttpClient      *http.Client
+	StreamingClient *http.Client
 
+	schema                    *graphql.Schema
 	subscriptionClientFactory graphqlDataSource.GraphQLSubscriptionClientFactory
 }
 
-func (p *proxyOnlyGraphQLEngineAdapter) EngineConfig() (*graphql.EngineV2Configuration, error) {
+func (p *ProxyOnly) EngineConfig() (*graphql.EngineV2Configuration, error) {
+	var err error
+	p.schema, err = parseSchema(p.ApiDefinition.GraphQL.Schema)
+	if err != nil {
+		return nil, err
+	}
+
 	staticHeaders := make(http.Header)
 
-	url := p.apiDefinition.Proxy.TargetURL
+	url := p.ApiDefinition.Proxy.TargetURL
 	if strings.HasPrefix(url, "tyk://") {
 		url = strings.ReplaceAll(url, "tyk://", "http://")
 		staticHeaders.Set(apidef.TykInternalApiHeader, "true")
@@ -31,15 +37,15 @@ func (p *proxyOnlyGraphQLEngineAdapter) EngineConfig() (*graphql.EngineV2Configu
 	upstreamConfig := graphql.ProxyUpstreamConfig{
 		URL:              url,
 		StaticHeaders:    staticHeaders,
-		SubscriptionType: graphqlSubscriptionType(p.apiDefinition.GraphQL.Proxy.SubscriptionType),
+		SubscriptionType: graphqlSubscriptionType(p.ApiDefinition.GraphQL.Proxy.SubscriptionType),
 	}
 
 	v2Config, err := graphql.NewProxyEngineConfigFactory(
 		p.schema,
 		upstreamConfig,
 		graphqlDataSource.NewBatchFactory(),
-		graphql.WithProxyHttpClient(p.httpClient),
-		graphql.WithProxyStreamingClient(p.streamingClient),
+		graphql.WithProxyHttpClient(p.HttpClient),
+		graphql.WithProxyStreamingClient(p.StreamingClient),
 		graphql.WithProxySubscriptionClientFactory(p.subscriptionClientFactory),
 	).EngineV2Configuration()
