+++ apidef/adapter/gqlengineadapter/utils_engine_v2.go
@@ -1,4 +1,4 @@
-package adapter
+package gqlengineadapter
 
 import (
 	"errors"
