+++ apidef/adapter/gqlengineadapter/utils_engine_v2_test.go
@@ -1,4 +1,4 @@
-package adapter
+package gqlengineadapter
 
 import (
 	"net/http"
